Request context for failover policy
Summary:
The FailoverRoute uses FailoverPolicy to determine the order in which the failover destinations are visited. To help policy make decisions, a per request context is necessary to aid this decision. For example, the number of retries is useful to help FailoverPolicy decide if it needs to reorder the destinations or skip destinations etc.

This change adds a FailoverPolicyContext.

Reviewed By: aptxkid

Differential Revision: D19375039

fbshipit-source-id: c2d0efb5589a32c5966bfc5010602e2380f06d93