Extend cross-region failover policy to skip regions without UDB
Summary:
To avoid failing over to regions without UDB replica, TAO can specify an optional list of regions that have UDB for the request shard in the request's carbon field "udbRegions". This optional list, when specified, is used to skip failover regions that are not on the list. If the optional list is not specified, no failover regions are skipped.

See https://fb.workplace.com/groups/602076086642106/permalink/1360337977482576/ for more details

Differential Revision: D20552053

fbshipit-source-id: 9a5f6c6c360fcffb428170953185c878a23e2dd8