Move assert check in cross region routing policy
Summary:
The simple cross region failover routes request to a region based on region names in -
1. Failover region names list
2. The master region in the request

At mcrouter configuration time, we make sure that failover region names have corresponding routes for #1. However #2 could be a bogus region name or a region name that is valid but has no route (as is the case for tnightly). In that case we were asserting too early, when checking if master region has a route. This change moves the assert to later after we've checked that master region has a route and marked it visited if its a bad value for which there's no route.

Differential Revision: D20671368

fbshipit-source-id: 5cfcdd9322af8ccbafdd55b3de06bb08b357b02f