Support for fail-open behavior in mcrouter
Summary:
Support for fail-open behavior in mcrouter

A pool level tko tracking config is supported. The config looks like
  "pools" {
    "A" : {
      ....
      "tko_tracker" : {
        "num_soft_tko_threshold_upper": 10,
        "num_soft_tko_threshold_lower": 7
      }
    }
  }

      OR

  "pools" {
    "A" : {
      ....
      "tko_tracker" : {
        "percent_soft_tko_threshold_upper": 50,
        "percent_soft_tko_threshold_lower": 35
      }
    }
  }

This fail-open behavior stops declaring unresponsive (timedout) servers in a
pool as soft-tko if the number of soft tkos in a pool reach
"num_soft_tko_threshold_upper" or "percent_soft_tko_threshold_upper".
This is considered as entering fail-open state.
And when the number of soft tkos in the pool fall to "num_soft_tko_threshold_lower" or
"percent_soft_tko_threshold_lower" fail-open state is exited.

Reviewed By: stuclar

Differential Revision: D20821600

fbshipit-source-id: 94750ed875fd58dbae1d6d30eb5217446d727ef7