External stats support in Standalone mcrouter
Summary:
This allows us to add stats to Standalone mcrouter, such that it can be invoked as:
```
echo stats standalone | nc <mcrouter listening port>
```

Standalone stats can be added on the router, by adding a callback:
```
router->standaloneStatsHandler().registerExternalStats(<namespace>, <callback>)
```

The callback is a function which takes in a reference to a `ExternalStatsData` object, which can then be populated.

The callback is invoked every time there is a user initiated request for stats, and is expected to be stateless and thread safe.

Reviewed By: stuclar

Differential Revision: D21314116

fbshipit-source-id: 1be3888f27023159301d9e7e4d5e892c451e2598