Migrate Memcache and McRouter internal IDLs using thrift namespace to _ref API only
Summary:
Adds support for a new option in the IDL "
gen_ref_api_only;", which indicates that to the compiler to generate the accessors using the field_ref API.

It is not possible to do this by default for all IDLs with a thrift namespace because due to Common.idl which is included by all services. Hence the need to do this in a more controlled manner.

Reviewed By: edenzik

Differential Revision: D21758508

fbshipit-source-id: 8131e52f908ca5d0666e33db18bbac2187f65fc5