Rendezvous Hash based failover policy
Summary:
In FNA, we often have hosts, or processes, that are down for a long
time.  When we add or delete hosts from the config, it's important that we
redirect the fewest possible requests.  In particular, when hosts have been
disabled for a while, and we then remove them from the config, it's important
that all keys continue to go to the same hosts.  Thus, the failover policy, when
machine X is down, needs to match the primary policy, when machine X is removed.
Since the primary policy is RendezvousHash, this failover policy uses the
RendezvousHash logic to choose the failover.

Reviewed By: ftimyo

Differential Revision: D20142226

fbshipit-source-id: 7385d03222aeb1b904db2f4e0d7c11f659992a97