Redo D20142226, which was reverted in D22305711.
Summary:
In FNA, we often have hosts, or processes, that are down for a long
time.  When we add or delete hosts from the config, it's important that we
redirect the fewest possible requests.  In particular, when hosts have been
disabled for a while, and we then remove them from the config, it's important
that all keys continue to go to the same hosts.  Thus, the failover policy, when
machine X is down, needs to match the primary policy, when machine X is removed.
Since the primary policy is RendezvousHash, this failover policy uses the
RendezvousHash logic to choose the failover.

Reviewed By: ftimyo

Differential Revision: D22775059

fbshipit-source-id: a2d27b66c4da31b0baa5d045bf63a96b5bf177f2