Use std::atomic for counting_sem.
Summary:
The previous implementation uses non-atomic read (via volatile) on the semaphore value and therefore fails TSAN. This diff changes the semaphore value type to std::atomic.

There is one little "non-standard-ness" here in that we're assuming that the address of the atomic coincides with the address of the int itself, since we're passing the atomic directly to the futex syscalls.

Reviewed By: stuclar

Differential Revision: D22404835

fbshipit-source-id: 2c75ba8ec07628d7c4e9e51dc460db61b7e66001