Fix a TSAN-reported race condition in mcrouter TestClientServerUtil.
Summary: The bug is that the config that is closured around by reference (TestClientServerUtil.h:131) is called in a spawned thread, and TestServer::run function does not guarantee to return after that lambda is already called. Instead, it uses an 100ms sleep. This diff changes the sleep to a baton.

Reviewed By: stuclar

Differential Revision: D23123475

fbshipit-source-id: ee1c6f63a2276e5c67f3bac1e6c0d94c656b00e2