Fix a race condition between addStartupOpts and getStartupOpts between the initialization thread and the logger thread.
Summary:
Example stack trace: P137251365

The initialization thread sets the startup options later than the logger thread gets a chance to run. This makes that safe. (The alternative is to have the logger thread start later, but I'm not familiar enough with the code to make that kind of change.)

Reviewed By: stuclar

Differential Revision: D22860043

fbshipit-source-id: be7186c5f7d784c2dfeaac66c9f4ae00b167cc5a