CarbonProtocolWriter exposes writer for binary field length
Summary:
Users of customized serialization traits may encode their fields with
the binary wire type. For backward compatibility, binary fields must
start with the length of the encoded value. If the serialization trait
is optimized to avoid an intermediate IOBuf, there was previously no
convenient way to write the field length.

Reviewed By: ifcharming

Differential Revision: D25342126

fbshipit-source-id: 7355f833fc53b63f66bbfd2ecbc25828b68236f3