Introduce FbMcrouterTransportContext to generalize identity extraction
Summary:
When using Fizz we aren't properly extracting MultiIdentity certs in Mcrouter This was discovered when trying to integrate Granular ACLs in BigCache (Granular ACLs use connection level certs to extract accessors on which to do the auth).

Instead of add another special case in `finalizeTransport` that works for Fizz, this diff removes our dependency in Mcrouter on `MultiIdentityCertificate` and instead relays on the well supported `getClientIdentitiesFromTransport` to extract an `unordered_set` of Identities.

This is implemented by wrapping any transport with an `FbMcrouterTransportContext`, which at transport finalization time parses out identities from the `x509` payload of the cert (in a generic way). This approach works for Fizz, TLS 1.2, etc.

Mutual auth, Connection level ACLs, and Granular ACLs have all been changed in this diff to work with the change.

In the future, we can use this to also add other Fb-specific properties to a transport during its lifecycle in OSS.

Reviewed By: mingtaoy

Differential Revision: D25299757

fbshipit-source-id: 3911a2b51d49fef169a4690bd1874b750e109fb6