Add destination host logging for SRRoute
Summary:
As above.

This is a temporary workaround until we consolidate on a single interface for returning host information from SR. Presently we have IP + port + MSB via lastUsed host on the exception path and just IP in string format for the normal path via thrift headers.

This diff enables makes use of the IP when we are logging.

Reviewed By: ftimyo

Differential Revision: D27695143

fbshipit-source-id: 0eccb60582bb09222bfb76ba5422955ef9a64d99