Obtain a short-lived lock when trying to preallocate TKO trackers vector
Summary:
We occasionally run into a TSAN issue (reported [here](https://fburl.com/sanitizer_analysis/r8c2ad4m)) where:
- We try to get the TkoTracker stats from one thread
- While trying to remove a TkoTracker from the TkoTrackerMap from another thread.

Most of this is done under mutex. However, getting the size of the TkoTrackersMap happens out of lock which creates a R/W conflict if a TkoTracker is concurrently being added or removed.

This diff takes a lock when getting the size (but before preallocating) which is very short lived and is subsequently being taken for a much longer amount of time.

See T79657998 for more context.

Reviewed By: stuclar

Differential Revision: D27960773

fbshipit-source-id: 3c0413c00ffc9a260ceb80fd87795c1c7509e24c