Pass threadPrefix to EventBasePool ctor and don't rename mcrouter threads
Summary:
For mcrouter integration we do not want to rename the threads that mcrouter passes in when initializing SRClientFactory. However we still need the thread prefix to correctly track stats for the threads:

https://www.internalfb.com/intern/diffusion/FBS/browsefile/master/fbcode/servicerouter/common/EventBasePool.cpp?commit=5854d8d58e147fca7f0c385fc2e74bab3cb2cc7d&lines=241

This diff passes the thread prefix from libmcrouter to EventBasePool and does not rename the threads.

NOTE: for other modules in SR like SmcConfigProvider, HealthCheckManager, ServiceSelectorCache, we will still create a singleton EventBasePool with SREventBase-named threads and use those in these places.

Reviewed By: spalamarchuk

Differential Revision: D28034554

fbshipit-source-id: 32460caa157399b8285539d1cb6bc9819499974d