add stats of cat request for rollout
Summary:
We are starting E2E test and will do the rollout soon for cat token propagation. The traffic will be mainly on devserver, on-demand, intern, which is relatively small.

(1) we added metrics in mcrouter_requests which is highly sampled, not idea to report small traffic.
(2) Add to upsampled stats is fairly hot code, but also fairly cheap so probably OK to do. (especially use UNLIKELY)

Reviewed By: edenzik

Differential Revision: D28225404

fbshipit-source-id: e3f00b38ea4d7ce4a0b28109238c0033cac2e34f