Remove MultiIdentityCertificate in Mcrouter
Summary:
The concept of a `MultiIdentityCertificate` is being phased out (citation needed). It is simply a decoration that allows us to:
- Replace the x509 payload in a peer cert with a list of `Identity` objects, which are an FB-specific concept
- Still refer to certificates (and the transports containing them) generically, enabling them to be OSS.

Hence to get the identities in a connection we simply extract the certificate, cast it to a `MultiIdentityCertiticate`, and get the identities extracted.

This approach is being phased out in favor of a more generic way of extracting peer identities, either by:
1) Direct extraction from the `transport` using the x509 payload of the peer certificate on every request (using `getClientIdentitiesFromTransport`)
2) A `clientIdentityHook` which runs at connection establishment time to populate `Cpp2ConnContext` with a `void*` typed list of identities. These identities are extracted from the transport using the hook, by calling #1. They can then be extracted with `getClientIdentitiesFromCtx`.

This diff removes all instances of `MultiIdentityCertificate` from Mcrouter code, replacing it instead with #1 above. However, this suffers from two drawbacks:
- Higher CPU overhead. As we need to parse the identities on the fast path on every request in gACL
- Higher memory overhead. As we're keeping both the underlying identities and the x509 payload.

Hence in subsequent diffs we will:
1. Move to using only `getClientIdentitiesFromCtx` everywhere, which uses the pre-parsed Identities extracted at connection time (saving CPU)
2. Use `dropPeerCertificate()` on the `Transport` after the client identity hook is executed (saving memory).

We'll avoid landing this diff before doing the above.

Concurrently, we should remove all instances of MultiIdentityCertificate from SR. Discussing this with AjanthanAsogamoorthy.

Reviewed By: stuclar

Differential Revision: D27926976

fbshipit-source-id: 44d8b2a31abd3e685e1956221797ad6f44ac471f