mcrouter: set tls_ticket_key_cb only if ticketKeys are provided in config
Summary:
**Problem**

When AsyncMcServer::Options.tlsTicketKeySeedPath is set to empty, SSL session ticket callbacks are expected to be disabled. However, session ticket callbacks are still enabled and this results in misleading logs on the server such as "No TLS ticket key available for encryption.".

**Root Cause**

SSLContextProvider expects a folly::optional<TicketKeySeeds> when creating server context. However, McServer blindly passes the TicketKeySeeds even when it is empty. Also, when creating the SSLServerContext, the SSLContextProvider does not check if the provided ticketKeySeeds are empty.

**Fix**

Fixed McServer connection accept callback to pass optional when ticketKeySeeds are empty. Also fixed createServerContext to only set tls_ticket_key_cb when ticketKeys are non-empty.

Reviewed By: matheweis, edenzik, ifcharming

Differential Revision: D30296683

fbshipit-source-id: 17a4fceedc6f6b47e943b01fef7bdd7225312661