Avoid copies when constructing ProxyConfig
Summary:
HHVM spends 0.1-0.2% gCPU here copying JSON config objects. About 2/3 of that time is in copying "named_handles" configs, and the other 1/3 is poolJson.

It's easy to fix poolJson - it's a reference into another JSON object, so changing auto -> auto& is sufficient.

The named handles require a bit more care. RouteHandleFactor::addNamed takes them by value, and modifies them to set the "name" field if it's not already set. Luckily, there's only one caller of addNamed - ProxyConfig - and only one caller of that - ProxyConfigBuilder. We move the "name" mutation to ProxyConfigBuilder, and make addNamed keep references to the (now const) configs, instead of their values.

Reviewed By: colavitam

Differential Revision: D31840394

fbshipit-source-id: 133cbd0898e30849da83778df7d2526bcf332cac