Don't rehash / reallocate in ShardSplitter
Summary:
We are inserting shard splits in a loop without pre-allocating (reserving) the size of the underlying F14Map. This incurs in reallocations and rehashes every time we run out of capacity in the map.

This diff allocates enough space ahead of time to prevent said wastage in CPU cycles (see details in the FB-specific part of the commit message).

No every iteration of the loop will insert elements in the hash table and this could cause some memory to be unused, so we would ideally call some `shrink_to_fit` method after we add all the elements. However, this method doesn't seem to exist in f14, and, in the average case, it won't cause more unused memory than now, as the automatic growth of this hash-table is exponential.

Reviewed By: marksantaniello

Differential Revision: D32021674

fbshipit-source-id: 99c240f24453aa906330b50a5e817afcecb0d734