Add ODS counter to track number of reconfig updates are skipped due to same version
Summary:
There is a version check in `tierUpdateCallback` that skips updates if the version number is the same.

This was added as an optimization to avoid unnecessary reconfigurations but is causing routing problems.

This diff introduces a counter that gets bumped every time we skip due to a version match.

Reviewed By: edenzik, stuclar

Differential Revision: D32175545

fbshipit-source-id: 55be7fa7b7f86ad5ab7158d1f8a4705fa5085423