docs: add GH button in support of Ukraine (#389)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

Pull Request resolved: https://github.com/facebook/mcrouter/pull/389

Reviewed By: disylh

Differential Revision: D34630826

Pulled By: dmitryvinn-fb

fbshipit-source-id: dbced0824e23153598ab4212d4c12b7806c6800c