Hash failure domain string for stable identifier between Mcrouter and ServiceRouter
Summary:
Currently MSB aware failovers are broken with SR + MCR. The identifier used on the SR side is the trailing in in the failure domain string, ie FB|PRN1|MSB_1. In mcrouter, its a static integer (when configured as a string).

This diff applies a SpookyHash to the string and makes the iface available to SR as a library so they can have the same view of failure domains.

Reviewed By: ftimyo

Differential Revision: D34917782

fbshipit-source-id: 081122bdaeab5beae62777fce4d91fefee1dd6f2