Return folly::None to Petra client when deserialization fails
Summary:
As a fix to HHVM DBG crashes, I've added a code that caught an exception to the deserialization in CarbonLookasideRoute (see D35987055 (https://github.com/facebook/mcrouter/commit/29d7aaa400b5eabfc1e23a2fecfd43f9c964547e)). This exception is not new - it has been throwing for quite some time before Jan 2022 but wasn't noticed because it was only failing in DBG (due to a LOG(DFATAL) mode for specific Petra keys.

As a result of the fix, Petra noticed an increase in error rate because in the event of this exception, we returned a LOCAL_ERROR to tell Petra to fetch from the backend.

In this diff, I modified the return error to instead return a folly::None to signal to the client to fetch from the backend without increasing error rate.

Differential Revision: D36228612

fbshipit-source-id: 2a8128eef4191a17d23edd25f2ff8724c7d6fddf