Return normalReply if there are failover destinations but we hit failoverPolicy maxErrorTries limit
Summary:
If there is only 1 child in failover route, when that fails we return the original error.

If there are more children to failover to but we hit failoverPolicy_.maxErrorTries that prevent failover then we return an empty reply because failoverReply isn't correctly initialized.

This fixes that by keep the behavior consistent with the case where there are no children i.e. return the original reply.

Reviewed By: alikhtarov, vrishal

Differential Revision: D36617004

fbshipit-source-id: 928367482794f08bb0c8b31e55a8dd316ddea281