Add test to check cache client and standalone mcrouter compatibility
Summary: S273625 was a result of a cacheclient change that used a protocol change that the standalone mcrouter did not have. This diff adds a test that uses taocmd (which uses the current cacheclient version on requests it sends) to send FbobjectGetRequest to a standalone mcrouter (the current release specified here https://fburl.com/code/nug583jv) that has a ClientCompatibilityCheckerRoute in its routing config. If the standalone mcrouter release has stale protocol, this test will fail since ClientCompatibilityCheckerRoute returns an error.

Reviewed By: anidev

Differential Revision: D37583320

fbshipit-source-id: bd5c5f867aef2cd820a8a33f8be2a7a1fead26dd