Add RoutableReplies to RouterInfo
Summary:
Add routable replies list to RouterInfo structure to allow SFINAE to filter out replies that don't contain the ReplyCommon mixin. This is needed to add a replySource bitmask to replies for Memcache/Ucache.

Note for all other services, this is a noop, its codegen that for now will not be used.

Reviewed By: alikhtarov

Differential Revision: D41239758

fbshipit-source-id: 626a1216d6d954241bd2c003f6fae1743b18c306