Avoid unnecessary copy of res in the lambda of ConfigPreprocessor::parseMacroDef
Summary:
This opportunity was found with the combination of [Infer's PULSE_UNNECESSARY_COPY analysis and Strobelight cycle counts](https://fb.workplace.com/groups/infer.eng/permalink/970566283876913/).

This shows up as one of the top issues ordered by inclusive cycle counts  with  **13.65T** cpu cycles in a day. This is a lot!

{F818628797}

Reviewed By: ot

Differential Revision: D41842397

fbshipit-source-id: 064151eff9939d6284fe6cf4c6211528a862ecfa