add BucketHashSelector
Summary:
The previous diff with this change D44109030 was reverted but we believe
special hash selector that does the selection based on a bucket id in the fiber context.
IMPORTANT: the bucket id in the fiber context must be present otherwise the selection throws. This means, the bucketized PoolRoute must be wrapped in McBucketRoute that computes the bucket id and adds it into the fiber context.

Next diff: adding bucketId to a request in `DestinationRoute`

Reviewed By: stuclar

Differential Revision: D44180783

fbshipit-source-id: e0b72f4cc3c941fe1a99c581179506512df9a552