Match Proxy EventBase to current EventBase in SameThread mode.
Summary:
Similar to the previous diff in this stack, we'd like to run proxygen workers on the same thread pool as mcrouter threads. The motivation for this is that it [significantly improves performance](https://fb.workplace.com/groups/proxygendevelopers/permalink/3023482191292485/) by reducing context switches and improving locality. Both mcrouter and proxygen worker threads run similar workloads in that they're both latency sensitive and non-blocking.

Presently when using the same thread mode, the router makes no attempt to set the proxy index and relies on the user doing this. Instead of relying on the user to set the proxy index, automatically find the corresponding proxy for the current EventBase and use that.

This requires that the caller create the client on the desired EventBase that will be used.

Reviewed By: stuclar, mjoras

Differential Revision: D44722434

fbshipit-source-id: 9875bb666eb31bb389285965db3c235409444017