lower log level from ERROR to WARNING for legacy setup path
Summary:
I recently changed ensureConfigDirectoryExists behavior in D47081565.
One of the changes was to log when the parent directory of the call doesn't exist.
However it's logged as ERROR.  That's good when this fails unexpectedly,
but do we have legacy use cases where this call is expected to fail silently?

Since we care about our use case, but others may now, how about
logging this the first time as WARNING.  We'll look for it in our releases
and for debugging customer deployments which require this behavior.
For all other teams, a single WARNING in logs should not trigger any
"regression" detection or other spurious error checking.

Reviewed By: udippant

Differential Revision: D47349146

fbshipit-source-id: 71cf19071c00d19ebf66e8fb4d14aeb76806411a