Revert D49245470: Multisect successfully blamed "D49245470: [mcrouter] Add core_multiplier_threshold option" for test or build failures
Summary:
This diff is reverting D49245470
D49245470: [mcrouter] Add core_multiplier_threshold option by stuclar has been identified to be causing the following test or build failures:

Tests affected:
- [mcrouter:mcpiper_integration_tests - test_delete_caret_raw (mcrouter.test.test_mcpiper.TestMcpiper)](https://www.internalfb.com/intern/test/281475056782537/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3050351
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: stuclar

Differential Revision: D49316247

fbshipit-source-id: 5916395d3eeae6438e9986828d42a3f0396661cb