stopTLS leverages ExtractedIdentityCert to keep credentials after dropping certificate
Summary:
When dropping down to plain, we currently keep the X509 payload in order to eventually extract identities.

This was attempted to be fixed in D49045451 - however that caused the identities to not be kept and authentication failure ensued.

This diff attempts to:
- Extract identities when dropping down to plain
- Drop the SSL context as soon as we do so

This ensures that we have identities downstream, without needing the X509 payload.

Reviewed By: frqiu

Differential Revision: D49611022

fbshipit-source-id: 7f4246c938c613dd13c973cb07300b53769f36df