strip routing prefix in all cases if configured
Summary:
The routing prefix should always be stripped.
We discovered samples where the routing prefix is not stripped; after investigations these are apparently only TKOs with PoolRoute: https://fburl.com/scuba/mcrouter_requests/6tq7mwtp
This is due to inconsistent logging code in the DestinationRoute and this diff fixes it.

Reviewed By: stuclar

Differential Revision: D50704110

fbshipit-source-id: d0e0a22c4537f52048320c40f928957124f4ee05