change log errors to warning for the best effort stats and backup directories
Summary: We get erroneous reports from users who inadvertently suspect these error log lines are fatal and causing crashes or other issues.  These are benign and tends to mean that mcrouter cannot dump config backups and stats to disk which is generally not a showstopper but can pose a reliability risk (thus warning).

Reviewed By: lenar-f

Differential Revision: D48568589

fbshipit-source-id: c955b39826a385a9a29e68f262783d1e0c92a528