Move initialization of isProxyThread_ to the thread factory
Summary: `isProxyThread_` is used for diagnostic purposes, to assert that some logic doesn't run in the proxy threads. We can enable it for all proxy threads by using the thread factory, so it is guaranteed that it will be set on all proxy threads even when using `MuxIOThreadPoolExecutor` (which may run the eventbases on a smaller set of threads).

Reviewed By: stuclar

Differential Revision: D54416135

fbshipit-source-id: 91512648d26cc90570055f9a9d4c2249f4a3a475