Make async_tm and async_eb HandlerCallback param a shared_ptr
Summary:
This diff does two things:
  1. Make the parameter of `async_tm_*` and `async_eb_*` callbacks in `ServiceHandler<>` implementations be a `shared_ptr` of `HandlerCallback<...>` instead of `unique_ptr`.
  2. Hide this detail behind an alias, `HandlerCallback<...>::Ptr` (which expands to `std::shared_ptr<HandlerCallback<...>>` and codemod all of fbcode.

(2) is useful independently of (1) but it seemed like a good opportunity to bundle both into one codemod.

## Why do we need `shared_ptr`?

Consider that:
1. `HandlerCallback`'s lifetime is tied to the lifetime of the Thrift's internal request object. That is, when `HandlerCallback` is destroyed, we mark the request as "done processing" and the schedule a task on the IO thread to deallocate the associated memory.
2. We ask the user to manage this lifetime, since we provide them a `unique_ptr<HandlerCallback>`. In other words, if they leak it then we leak the request-internal memory as well. This is not a problem usually because the user just calls `callback->{result,exception,done}` and then lets the object go out of scope.

Now with `ServiceInterceptor`, we want to be able to inject an event handling customization point *before* deserialization and we want it to be asynchronous. But `HandlerCallback::{result,exception,done}` are all synchronous. Furthermore, `result` and `exception` delegate to user-overridable `doResult` and `doException` functions!

This means that the only reliable place where we can invoke `ServiceInterceptor::onResponse` is in `result` / `exception` functions themselves.

We can't simply make these functions asynchronous / as coroutines — that's a very big breaking change that would require a lot of code changes across fbcode.

We also can't block the thread inside the implementation of `result` / `exception` because... well it would take down our infra.

Unfortunately, we also cannot **asynchronously schedule invocation of `ServiceInterceptor::onResponse`** before calling the user-defined `doResult`. Why? Because, by that time, the `unique_ptr` would have gone out of scope and thus the request would be deleted.

Here lies the trick — instead of giving the user *exclusive* ownership of `HandlerCallback`, we provide *shared* ownership with the runtime code. In other words, if there are `ServiceInterceptor`s installed, then the runtime will extend the lifetime of `HandlerCallback` (and thus the request) while they are executing, even after the callback object provided to user code has gone out of scope.

This is how `HandlerCallback<>::result` might be implemented:
```

  void result(InputType r) {
#if FOLLY_HAS_COROUTINES
    if (!hasServiceInterceptorsInstalled()) {
      doResult(std::forward<InputType>(r));
    } else {
      this->processServiceInterceptorsOnResponse()
          .semi()
          .deferValue(
              [callback = shared_from_this()),
               result = std::decay_t<InputType>(static_cast<InputType>(r))](
                  auto&&) mutable { callback->doResult(std::move(result)); })
          .via(executor_);
    }
#else
    doResult(std::forward<InputType>(r));
#endif // FOLLY_HAS_COROUTINES
  }
```

And in fact, I've verified that it works: D55459493.

Calling `folly::to_shared_ptr` on the callback seems to be a relatively common pattern in user code anyways for non-trivial cases.

## Performance?

In practice we are not doing much more work. In the case that no `ServiceInterceptor` is installed, the refcount always stays at 1. So the added work is an (uncontested) atomic deref of the refcount. Note that since we use `make_shared` in generated code, we are not adding any mallocs.

If a `ServiceInterceptor` is installed, then the execution of the `ServiceInterceptor` implementation will dominate the execution time.

 ---

Reviewed By: sethdelliott

Differential Revision: D55506070

fbshipit-source-id: 4feec3b1b39387856019c7c10b6b1a1a2b704d1a